% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_add_dependency.R
\name{ec_add_dependency}
\alias{ec_add_dependency}
\title{Add plugin dependencies to echarts objects}
\usage{
ec_add_dependency(ec, name = NULL)
}
\arguments{
\item{ec}{A `echarter` `htmlwidget` object.}

\item{name}{The partial name or js file}
}
\description{
Add plugin dependencies to echarts objects
}
\examples{

}
