% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_add_series.R, R/echarter.R
\name{ec_add_series}
\alias{ec_add_series}
\alias{ec_add_series}
\alias{ec_add_series.default}
\alias{ec_add_series}
\alias{ec_add_series.data.frame}
\alias{ec_add_series}
\alias{ec_add_series.matrix}
\alias{ec_add_series}
\alias{ec_add_series.numeric}
\alias{ec_add_series}
\alias{ec_add_series.character}
\alias{ec_add_series}
\alias{ec_add_series.ts}
\alias{ec_add_series}
\alias{ec_add_series.forecast}
\alias{ec_add_series}
\alias{echarter.data.frame}
\title{Adding series from echarts objects}
\usage{
ec_add_series(ec, data = NULL, ...)

\method{ec_add_series}{default}(ec, ...)

\method{ec_add_series}{data.frame}(ec, data, mapping = ecaes(), ...)

\method{ec_add_series}{matrix}(ec, data, ...)

\method{ec_add_series}{numeric}(ec, data, ...)

\method{ec_add_series}{character}(ec, data, ...)

\method{ec_add_series}{ts}(ec, data, ...)

\method{ec_add_series}{forecast}(ec, data, addOriginal = TRUE,
  addLevels = TRUE, fillOpacity = 0.1, name = NULL, ...)

\method{echarter}{data.frame}(data, mapping = ecaes(), ...,
  theme = "default", width = NULL, height = NULL, elementId = NULL,
  dispose = TRUE, renderer = "canvas")
}
\arguments{
\item{ec}{A `echarter` object.}

\item{data}{An R object like list, data.frame, matrix, numeric, character, ts, forecast, etc.}

\item{...}{Arguments defined in \url{http://echarts.baidu.com/option.html#series}.}

\item{mapping}{a ecaes object}

\item{addOriginal}{Logical value to add the original series or not.}

\item{addLevels}{Logical value to show predictions bands.}

\item{fillOpacity}{The opacity of bands.}

\item{name}{The name of the series.}

\item{mapping}{a ecaes object}
}
\description{
Adding series from echarts objects

`ec_add_series.data.frame` for data.frame objects

`ec_add_series.matrix` for matrix objects

`ec_add_series.numeric` for numeric objects

`ec_add_series.character` for character objects

`ec_add_series.ts` for ts objects

`ec_add_series.forecast` for numeric objects

`echarter.data.frame` for data.frame objects
}
\examples{

weekDays <- c('Mon','Tues','Wed','Thurs','Fri','Sat','Sun')
dat <- data.frame(
  saleNum = round(runif(21, 20, 100), 0),
  fruit = c(rep("Apple", 7), rep("Pear", 7), rep("Banana", 7)),
  weekDay = c(rep(weekDays,3)),
  price = round(runif(21, 10, 20), 0),
  stringsAsFactors = FALSE)

mapping = ecaes(x = weekDay, y = saleNum, group = fruit)

echart() \%>\%
  ec_add_series(
    data = dat, type = 'bar',
    mapping = ecaes(x = weekDay, y = saleNum, group = fruit))


weekDays <- c('Mon','Tues','Wed','Thurs','Fri','Sat','Sun')
dat <- data.frame(
  saleNum = round(runif(21, 20, 100), 0),
  fruit = c(rep("Apple", 7), rep("Pear", 7), rep("Banana", 7)),
  weekDay = c(rep(weekDays,3)),
  price = round(runif(21, 10, 20), 0),
  stringsAsFactors = FALSE)

echarter(
  data = dat, type = 'bar',
  mapping = ecaes(x = weekDay, y = saleNum, group = fruit))

}
