% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ec_add_theme}
\alias{ec_add_theme}
\title{Add echarts theme}
\usage{
ec_add_theme(ec, theme, series = TRUE)
}
\arguments{
\item{theme}{A `ec_theme` object containing options defined as
\url{http://echarts.baidu.com/option.html}.}
}
\description{
Add echarts theme
}
