% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_angleAxis}
\alias{ec_angleAxis}
\title{angleAxis}
\usage{
ec_angleAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the angleAxis.
(\url{https://echarts.apache.org/zh/option.html#angleAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
The angle axis in Polar Coordinate.
}
