% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_aria}
\alias{ec_aria}
\title{aria}
\usage{
ec_aria(ec, ..., baseoption = FALSE, add = TRUE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the aria
(\url{https://echarts.apache.org/zh/option.html#aria}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
generating description for charts automatically.
}
