% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_brush}
\alias{ec_brush}
\title{brush}
\usage{
ec_brush(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the brush
(\url{http://www.echartsjs.com/option.html#brush}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
brush component.
}
