% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_calendar}
\alias{ec_calendar}
\title{calendar}
\usage{
ec_calendar(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the calendar
(\url{http://www.echartsjs.com/option.html#calendar}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Calendar Coordinates
}
