% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ec_colorize}
\alias{ec_colorize}
\title{Create vector of color from vector}
\usage{
ec_colorize(x, colors = c("#c23531", "#2f4554", "#61a0a8", "#d48265",
  "#91c7ae", "#749f83", "#ca8622", "#bda29a", "#6e7074", "#546570", "#c4ccd3"))
}
\arguments{
\item{x}{A numeric, character or factor object.}

\item{colors}{A character string of colors (ordered) to colorize `x`}
}
\description{
Create vector of color from vector
}
\examples{

ec_colorize(runif(10))

ec_colorize(LETTERS[rbinom(20, 5, 0.5)], c("#FF0000", "#00FFFF"))

}
