% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_dataset}
\alias{ec_dataset}
\title{dataset}
\usage{
ec_dataset(ec, data, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{data}{An R object like json, data.frame, matrix.}

\item{...}{Additional arguments for the dataset
(\url{https://echarts.apache.org/zh/option.html#dataset}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
dataset for echarts, brings convenience in data management
}
