% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_dataZoom}
\alias{ec_dataZoom}
\title{dataZoom}
\usage{
ec_dataZoom(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the dataZoom
(\url{http://www.echartsjs.com/option.html#dataZoom}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
dataZoom component.
}
