% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_geo}
\alias{ec_geo}
\title{geo}
\usage{
ec_geo(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the geo
(\url{http://www.echartsjs.com/option.html#geo}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Geographic Coordinates
}
