% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_get_opt}
\alias{ec_get_opt}
\title{Get echarter options}
\usage{
ec_get_opt(ec, opt = NULL, index = NULL, num = NULL)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{opt}{Arguments, eg:series.type or series2.data}

\item{index}{The id of opt, if set \code{opt = series2, index = 0}, then select the opt by id.}

\item{num}{The num of opt. if opt = series2, then num = 2.}
}
\description{
Get echarter options
}
