% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_graphic}
\alias{ec_graphic}
\title{graphic}
\usage{
ec_graphic(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the graphic
(\url{http://www.echartsjs.com/option.html#graphic}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
graphic component.
}
