% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_grid}
\alias{ec_grid}
\title{grid}
\usage{
ec_grid(ec, ..., baseoption = FALSE, add = TRUE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the grid
(\url{https://echarts.apache.org/zh/option.html#grid}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Rectangular Coordinate
}
