% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ec_list_parse}
\alias{ec_list_parse}
\alias{ec_list_parse_}
\title{Convert an object to list with identical structure}
\usage{
ec_list_parse(df)

ec_list_parse_(df)
}
\arguments{
\item{df}{A data frame to parse to list}
}
\description{
\code{ec_list_parse} are similar to \code{rlist::list.parse} but this removes names.

\code{ec_list_parse_} set result name of ec_list_parse is null
}
\examples{

x <- data.frame(a=1:3, type=c('A','C','B'), stringsAsFactors = FALSE)
ec_list_parse(x)
ec_list_parse_(x)

}
