% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_mapbox3D}
\alias{ec_mapbox3D}
\title{mapbox3D}
\usage{
ec_mapbox3D(ec, ...)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the aria
(\url{https://echarts.apache.org/en/option-gl.html#mapbox3D}).}
}
\description{
The component can draw a 3D GeoJSON and coordinate system. Developers can display 3D Scatter, 3D Bubble, 3D Bar, 3D Lines on it.
}
