% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_mark}
\alias{ec_mark}
\alias{ec_markPoint}
\alias{ec_markLine}
\alias{ec_markArea}
\title{mark}
\usage{
ec_mark(ec, ..., markname = "markPoint", serie = NULL)

ec_markPoint(ec, ..., serie = NULL)

ec_markLine(ec, ..., serie = NULL)

ec_markArea(ec, ..., serie = NULL)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the mark.
markPoint see (\url{https://echarts.apache.org/zh/option.html#series-line.markPoint}).
markLine see (\url{https://echarts.apache.org/zh/option.html#series-line.markLine}).
markArea see (\url{https://echarts.apache.org/zh/option.html#series-line.markArea}).}

\item{markname}{The mark name, markPoint, markLine, markArea}

\item{serie}{The serie index of ec}
}
\description{
Mark an point, line, area in echarts
}
