% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_media}
\alias{ec_media}
\title{media}
\usage{
ec_media(ec, query, option, baseoption = FALSE, add = TRUE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{option}{default FALSE}

\item{add}{default TRUE}
}
\description{
media component for responsive.
}
