% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_options}
\alias{ec_options}
\title{options}
\usage{
ec_options(ec, ecs)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{ecs}{An \code{echarter} object list.}
}
\description{
echarts options
}
