% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_parallel}
\alias{ec_parallel}
\title{parallel}
\usage{
ec_parallel(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the parallel
(\url{https://echarts.apache.org/zh/option.html#parallel}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Parallel Coordinates
}
