% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_parallelAxis}
\alias{ec_parallelAxis}
\title{parallelAxis}
\usage{
ec_parallelAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the parallelAxis
(\url{http://www.echartsjs.com/option.html#parallelAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
the coordinate axis for Parallel Coordinates.
}
