% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_radar}
\alias{ec_radar}
\title{radar}
\usage{
ec_radar(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the radar
(\url{http://www.echartsjs.com/option.html#radar}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Radar Coordinate.
}
