% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_radiusAxis}
\alias{ec_radiusAxis}
\title{radiusAxis}
\usage{
ec_radiusAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the radiusAxis.
(\url{http://www.echartsjs.com/option.html#radiusAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
The Radial axis in Polar Coordinate.
}
