% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_registerMap}
\alias{ec_registerMap}
\title{Register map}
\usage{
ec_registerMap(ec, mapName, geoJson, specialAreas = NULL)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{mapName}{Name of map}

\item{geoJson}{\href{Geojson}{http://geojson.org/}.}

\item{specialAreas}{}
}
\description{
Register a \href{geojson}{http://geojson.org/} map.
}
\examples{
USA_geoJson <- jsonlite::read_json("http://www.echartsjs.com/gallery/data/asset/geo/USA.json")

USArrests_ <- USArrests \%>\%
  dplyr::mutate(states = row.names(.))

echart() \%>\%
  ec_registerMap("USA", USA_geoJson) \%>\%
  ec_add_series(
    type = 'map', mapType = 'USA',
    data = USArrests_,
    mapping = ecaes(name = states, value = Murder)) \%>\%
  ec_visualMap(
    calculable = TRUE,
    min = 0, max = 20, text = c("高", "低"),
    color = c('#d94e5d','#eac736')) \%>\%
  ec_tooltip(trigger = 'item',formatter = '{b}: {c}')

shanghai_geoJson <- jsonlite::read_json("https://raw.githubusercontent.com/ecomfe/echarts-www/master/asset/map/json/province/shanghai.json")

shanghai_dat <- data.frame(name = c("崇明区","静安区","宝山区","嘉定区","青浦区","虹口区","杨浦区","黄浦区","卢湾区","长宁区","浦东新区","松江区","金山区","奉贤区","普陀区","闵行区","徐汇区"), n = round(runif(17,1,100),0), stringsAsFactors = FALSE)

echart() \%>\%
 ec_registerMap("shanghai", shanghai_geoJson) \%>\%
 ec_add_series(
   type = 'map', mapType = 'shanghai',
   data = shanghai_dat,
   mapping = ecaes(name = name, value = n),
   label = list(
     normal = list(show = FALSE),
     emphasis = list(show = FALSE))) \%>\%
 ec_visualMap(
   calculable = TRUE,
   min = 0, max = 100, text = c("高", "低"),
   color = c('#d94e5d','#eac736')) \%>\%
 ec_tooltip(trigger = 'item',formatter = '{b}: {c}')

}
