% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_series}
\alias{ec_series}
\title{series}
\usage{
ec_series(ec, ..., baseoption = FALSE, add = TRUE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the series
(\url{https://echarts.apache.org/zh/option.html#series}).}
}
\description{
series.
}
\examples{
library(dplyr)
library(echarter)
weekDays <- c('Mon','Tues','Wed','Thurs','Fri','Sat','Sun')
echart() \%>\%
  ec_xAxis(type = 'category', data = weekDays) \%>\%
  ec_yAxis(type = 'value') \%>\%
  ec_series(
    type = 'line',
    name = 'Apple',
    data = as.integer(runif(7, 20,100)))

echart() \%>\%
  ec_xAxis(type = 'category', data = weekDays) \%>\%
  ec_yAxis(type = 'value') \%>\%
  ec_add_series(
    type = 'line',
    name = 'Apple',
    data = as.integer(runif(7, 20,100)))

dat <- data.frame(
  saleNum = round(runif(21, 20, 100), 0),
  fruit = c(rep("Apple", 7), rep("Pear", 7), rep("Banana", 7)),
  weekDay = c(rep(weekDays,3)),
  price = round(runif(21, 10, 20), 0),
  stringsAsFactors = FALSE)

dat_sp <- dat \%>\%
  select(fruit, weekDay, saleNum) \%>\%
  tidyr::spread(fruit, saleNum) \%>\%
  arrange(match(weekDay, weekDays))

echart() \%>\%
  ec_xAxis(type = 'category', data = weekDays) \%>\%
  ec_yAxis(type = 'value') \%>\%
  ec_dataset(data = dat_sp) \%>\%
  ec_series(
    name = "Apple",
    datasetIndex = 0,
    type = 'line', encode = list(y = "Apple")) \%>\%
  ec_series(
    name = "Banana",
    datasetIndex = 0,
    type = 'line', encode = list(y = 2, tooltip = c(0, 3)))

}
