% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_singleAxis}
\alias{ec_singleAxis}
\title{singleAxis}
\usage{
ec_singleAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the singleAxis
(\url{http://www.echartsjs.com/option.html#singleAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
An axis with a single dimension
}
