% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ec_theme}
\alias{ec_theme}
\title{define ec_theme object}
\usage{
ec_theme(...)
}
\arguments{
\item{...}{A `list` object containing options defined as
\url{http://echarts.baidu.com/option.html}.}
}
\description{
define ec_theme object
}
