% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_timeline}
\alias{ec_timeline}
\title{timeline}
\usage{
ec_timeline(ec, ..., baseoption = TRUE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the timeline
(\url{https://echarts.apache.org/zh/option.html#timeline}).}

\item{baseoption}{default TRUE}

\item{add}{default FALSE}
}
\description{
timeline component.
}
