% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_timeline2}
\alias{ec_timeline2}
\title{timeline2}
\usage{
ec_timeline2(ecs, ...)
}
\arguments{
\item{ecs}{An \code{echarter} object list.}

\item{...}{Additional arguments for the timeline
(\url{https://echarts.apache.org/zh/option.html#timeline}).}
}
\description{
timeline component.
}
