% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_toolbox}
\alias{ec_toolbox}
\title{toolbox}
\usage{
ec_toolbox(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the toolbox
(\url{https://echarts.apache.org/zh/option.html#toolbox}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
toolbox component.
}
