% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_tooltip}
\alias{ec_tooltip}
\title{tooltip}
\usage{
ec_tooltip(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the tooltip
(\url{https://echarts.apache.org/zh/option.html#tooltip}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
tooltip component.
}
