% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_visualMap}
\alias{ec_visualMap}
\title{visualMap}
\usage{
ec_visualMap(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the visualMap
(\url{https://echarts.apache.org/zh/option.html#visualMap}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
visualMap component.
}
