% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_xAxis}
\alias{ec_xAxis}
\title{xAxis}
\usage{
ec_xAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the xAxis
(\url{http://www.echartsjs.com/option.html#xAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
The x axis in Rectangular Coordinate
}
