% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_yAxis}
\alias{ec_yAxis}
\title{yAxis}
\usage{
ec_yAxis(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the yAxis
(\url{https://echarts.apache.org/zh/option.html#yAxis}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
The y axis in Rectangular Coordinate
}
