% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_yAxis3D}
\alias{ec_yAxis3D}
\title{yAxis3D}
\usage{
ec_yAxis3D(ec, ...)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the aria
(\url{https://echarts.apache.org/en/option-gl.html#yAxis3D}).}
}
\description{
The Y-axis in a 3D cartesian coordinate system.
}
