% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echart.R
\name{echart}
\alias{echart}
\title{Create a Echarts chart widget}
\usage{
echart(
  opt = list(),
  theme = "default",
  width = NULL,
  height = NULL,
  elementId = NULL,
  dispose = TRUE,
  renderer = "canvas",
  ...
)
}
\arguments{
\item{opt}{A `list` object containing options defined as
\url{https://echarts.apache.org/zh/option.html}.}

\item{theme}{A \code{ec_theme} class object}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}

\item{dispose}{Set to \code{TRUE} to force redraw of chart, set to \code{FALSE} to update.}

\item{renderer}{Renderer, takes \code{canvas} (default) or \code{svg}.}

\item{...}{Any other argument.}
}
\description{
This function creates a Echarts chart using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
