% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarter.R
\name{echarter}
\alias{echarter}
\alias{echarter}
\alias{echarter.default}
\alias{echarter}
\alias{echarter.numeric}
\alias{echarter}
\alias{echarter.character}
\alias{echarter}
\alias{echarter.ts}
\alias{echarter}
\alias{echarter.forecast}
\title{Create a echarts object from a particular data type}
\usage{
echarter(data, ...)

\method{echarter}{default}(data, ...)

\method{echarter}{numeric}(data, ...)

\method{echarter}{character}(data, ...)

\method{echarter}{ts}(data, ...)

\method{echarter}{forecast}(data, addOriginal = TRUE, addLevels = TRUE,
  fillOpacity = 0.1, name = NULL, ...)
}
\arguments{
\item{data}{A data object.}

\item{...}{Additional arguments for the data series
(\url{http://echarts.baidu.com/option.html#series}).}

\item{addOriginal}{Logical value to add the original series or not.}

\item{addLevels}{Logical value to show predictions bands.}

\item{fillOpacity}{The opacity of bands.}

\item{name}{The name of the series.}
}
\description{
\code{echarter} uses `echart` to draw a particular plot for an
object of a particular class in a single command. This defines the S3
generic that other classes and packages can extend.

`echarter.numeric` for numeric objects

`echarter.character` for character objects

`echarter.ts` for ts objects

`echarter.forecast` for numeric objects
}
\details{
Run \code{methods(echart)} to see what objects are supported.
}
