% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echart.R
\name{echarts-shiny}
\alias{echarts-shiny}
\alias{echartsOutput}
\alias{echarts-shiny}
\alias{renderEcharts}
\title{Widget output function for use in Shiny}
\usage{
echartsOutput(outputId, width = "100\%", height = "400px")

renderEcharts(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{The name of the input.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{A echarts expression.}

\item{env}{A enviorment.}

\item{quoted}{A boolean value.}
}
\description{
Output and render functions for using echarts within Shiny
applications and interactive Rmd documents.

Widget render function for use in Shiny
}
\details{
Widget output function for use in Shiny
}
