% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echart.R
\name{is.echart}
\alias{is.echart}
\alias{is.echarter}
\title{Reports whether x is a echarts object}
\usage{
is.echart(x)

is.echarter(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a echarts object
}
