% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_grid3D}
\alias{ec_grid3D}
\title{grid3D}
\usage{
ec_grid3D(ec, ...)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the aria
(\url{https://echarts.apache.org/en/option-gl.html#grid3D}).}
}
\description{
3D cartesian coordinate system component. It requires \code{ec_xAxis3D}, \code{ec_yAxis3D} and \code{ec_zAxis3D} axis components to be used together.
}
