% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_polar}
\alias{ec_polar}
\title{polor}
\usage{
ec_polar(ec, ..., baseoption = FALSE, add = FALSE)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{...}{Additional arguments for the polor.
(\url{https://echarts.apache.org/zh/option.html#polar}).}

\item{baseoption}{default FALSE}

\item{add}{default FALSE}
}
\description{
Polar Coordinate.
}
