% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-ec.R
\name{ec_registerMap}
\alias{ec_registerMap}
\title{Register map}
\usage{
ec_registerMap(ec, mapName, geoJson, specialAreas = NULL)
}
\arguments{
\item{ec}{An \code{echarter} object as returned by \code{\link{echart}}.}

\item{mapName}{Name of map}

\item{geoJson}{\href{Geojson}{http://geojson.org/}.}

\item{specialAreas}{}
}
\description{
Register a \href{geojson}{http://geojson.org/} map.
}
\examples{
USA_geoJson <- jsonlite::read_json("http://www.echartsjs.com/gallery/data/asset/geo/USA.json")

USArrests_ <- USArrests \%>\%
  dplyr::mutate(states = row.names(.))

echart() \%>\%
  ec_registerMap("USA", USA_geoJson) \%>\%
  ec_add_series(
    type = 'map', mapType = 'USA',
    data = USArrests_,
    mapping = ecaes(name = states, value = Murder)) \%>\%
  ec_visualMap(
    calculable = TRUE,
    min = 0, max = 20, text = c("高", "低"),
    color = c('#d94e5d','#eac736')) \%>\%
  ec_tooltip(trigger = 'item',formatter = '{b}: {c}')

}
