% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarter.R
\name{echarter}
\alias{echarter}
\alias{echarter.default}
\alias{echarter.numeric}
\alias{echarter.character}
\alias{echarter.ts}
\alias{echarter.forecast}
\title{Create a echarts object from a particular data type}
\usage{
echarter(data, ...)

\method{echarter}{default}(data, ...)

\method{echarter}{numeric}(
  data,
  ...,
  theme = "default",
  width = NULL,
  height = NULL,
  elementId = NULL,
  dispose = TRUE,
  renderer = "canvas"
)

\method{echarter}{character}(
  data,
  ...,
  theme = "default",
  width = NULL,
  height = NULL,
  elementId = NULL,
  dispose = TRUE,
  renderer = "canvas"
)

\method{echarter}{ts}(
  data,
  ...,
  theme = "default",
  width = NULL,
  height = NULL,
  elementId = NULL,
  dispose = TRUE,
  renderer = "canvas"
)

\method{echarter}{forecast}(
  data,
  addOriginal = TRUE,
  addLevels = TRUE,
  fillOpacity = 0.1,
  name = NULL,
  ...,
  theme = "default",
  width = NULL,
  height = NULL,
  elementId = NULL,
  dispose = TRUE,
  renderer = "canvas"
)
}
\arguments{
\item{data}{A data object.}

\item{...}{Additional arguments for the data series
(\url{https://echarts.apache.org/zh/option.html#series}).}

\item{addOriginal}{Logical value to add the original series or not.}

\item{addLevels}{Logical value to show predictions bands.}

\item{fillOpacity}{The opacity of bands.}

\item{name}{The name of the series.}
}
\description{
\code{echarter} uses `echart` to draw a particular plot for an
object of a particular class in a single command. This defines the S3
generic that other classes and packages can extend.
}
\details{
Run \code{methods(echart)} to see what objects are supported.
}
